requirements.txt
langchain
openai
pdf2image
pytesseract
python-dotenv
slack_sdk